fix: break remaining shared/mod.js barrel imports in report generation chain
PR #1527 fixed metrics.ts but missed several other paths that still
reach shared/mod.js → ui.js → @gsd/pi-tui during report generation
via native dynamic import() (which bypasses jiti alias resolution).

Remaining chains fixed:
- preferences.ts, preferences-validation.ts, export.ts, forensics.ts,
  migrate/parsers.ts: import from shared/format-utils.js directly
- state.ts, visualizer-data.ts, files.ts: import from milestone-ids.js
  instead of guided-flow.js (which pulls in shared/mod.js)
- files.ts: import checkExistingEnvKeys from new env-utils.ts instead
  of get-secrets-from-user.ts (which imports @gsd/pi-tui)

New file: env-utils.ts extracts the pure checkExistingEnvKeys function.
